Refactor to co-locate Dockerfiles with their applications
This removes the need to copy e.g. the build output for the frontend
into a directory under the Dockerfile for nginx so that it can be found
in the Docker build context.
Prior to this change there were to copies of e.g. the api server source
and the frontend build products, which risks unforced errors e.g.
accidentally copying over meaningful edits to the frontend source.